tweak(onesync): refine net time sync code
The goal of this commit is to improve the readability of `rage::netTimeSync` related code. Additionally, it aims to improve the maintainability of `CloneExperiments.cpp` by logically splitting certain parts of the file into appropriate chunks. The code refine isn't intended to be a full fledged refactor but rather a first step towards facilitating future refactoring efforts.
