+++ code/components/gta-net-rdr3/component.lua
@@ -17,6 +17,7 @@ return function()
 		'components/gta-net-five/src/CloneObjectManager.cpp',
 		'components/gta-net-five/src/ObjectIdManager.cpp',
 		'components/gta-net-five/src/rlNetBuffer.cpp',
+		'components/gta-net-five/src/netTimeSync.cpp',
 		'components/gta-net-five/src/MumbleVoice.cpp',
 	}
 end
